Added new header file containing the functions and macros from gni_priv needed to implement key translation. Additionally, added code to create new nic if one isn't available.
